Graph polish

+++ crypto/coin/query.py
@@ -16,7 +16,7 @@ class CoinController:
         self.urls   = []
 
         # TEMPORARY
-        self.path   = os.path.dirname(os.path.realpath(__file__)) + "\\"
+        self.path   = os.path.dirname(os.path.realpath(__file__)) + "/"
         
     def setCoin(self, url):
         if 'btc' in url:
@@ -42,7 +42,7 @@ class CoinController:
 
             query     = "MATCH (a:{})-[r:{}TX]->(b:{}) WHERE {} {}".format(self.coin, self.coin, self.coin, addrsText, query)
         elif params['sender[]'] and params['receiver[]']:
-            addrsText = "({}) AND".format(self.getAddrsText(params['sender[]'], params['receiver[]']))
+            addrsText = "({}) AND".format(self.getAddrsText(params['sender[]'], params['receiver[]'], rel=False))
 
             query     = "MATCH (a:{})-[r:{}TX]->(b:{}) WHERE {} {}".format(self.coin, self.coin, self.coin, addrsText, query.replace("WITH startNode(r) as a, endNode(r) as b, r ", ""))
         else:
@@ -62,9 +62,9 @@ class CoinController:
                 "to": bNode['id'],
                 #"id": rel['id'],
                 "value": float(rel['amount']),
-                "source": aNode['label'] or aNode['addr'],
+                "source": self.getKnown(aNode['addr'])['name'],
                 "sourceAddr": aNode['addr'],
-                "target": bNode['label'] or bNode['addr'],
+                "target": self.getKnown(bNode['addr'])['name'],
                 "targetAddr": bNode['addr'],
                 "type": rel['type'],
                 "amount": float(rel['amount']),
@@ -79,17 +79,17 @@ class CoinController:
 
         return edges
 
-    def getAddrInfo(self, addr, filters):    
-                    
+    def getAddrInfo(self, addr, filters):
+
         info = {
             "label": "",
             "addr": None,
             "totalTxs": 0,
-            "lastTx": 1230940800,
+            "lastTx": 1230940800, #REMOVE
             "balance": 0,
             "filters": {}
         }
-          
+
         if addr == "0":
             info["label"]    = "All Addresses"
         elif filters['addr[]']:
@@ -100,7 +100,7 @@ class CoinController:
 
             info['addr']     = filters['addr[]']
             
-            info['balance']  = self.driver.session().run("MATCH (a: " + self.coin + ") WHERE " + self.getAddrsText(info['addr']) + " RETURN SUM(a.balance)").single().value()
+            info['balance']  = self.driver.session().run("MATCH (a: " + self.coin + ") WHERE " + self.getAddrsText(info['addr'], rel=False) + " RETURN SUM(a.balance)").single().value()
         else:
             node = self.driver.session().run("MATCH (a:" + self.coin + ") WHERE a.addr = {addr} RETURN a", addr=addr).single()
             
@@ -142,18 +142,25 @@ class CoinController:
         
         addrsText = ''
 
-        if params['addr[]']:
-            addrsText = "({}) AND".format(self.getAddrsText(params['addr[]']))
-
-        query = "MATCH (a:{})-[r:{}TX]->(b:{}) WHERE {} {}".format(self.coin, self.coin, self.coin, addrsText, query)
-        
-        txs   = self.runFilters(query, params)
-        
         data = {
             'nodes': [],
             'edges': [],
             'totalTxs': 0
         }
+
+        if params['addr[]']:
+            """
+            for addr in params['addr[]']:
+                if not self.isValidAddr(addr):
+                    return data
+            """
+
+            addrsText = "({}) AND".format(self.getAddrsText(params['addr[]']))
+
+
+        query = "MATCH (a:{})-[r:{}TX]->(b:{}) WHERE {} {}".format(self.coin, self.coin, self.coin, addrsText, query)
+        
+        txs   = self.runFilters(query, params)
         
         id = 0
         for nodes in txs:
@@ -161,15 +168,16 @@ class CoinController:
             if nodes is None:
                 continue
             aNode = nodes['from']
+            aKnown = self.getKnown(aNode['addr'])
             aNode = {
                 "id": aNode.id,
-                "label": aNode['name'] or aNode['addr'],
+                "label": aKnown['name'],
                 "addr": aNode['addr'],
                 "balance": float(aNode['balance'] or 0),
                 "balVal": float(aNode['balance'] or 0),
                 "group": aNode['wallet'] or 'usdt',
                 "lastUpdate": aNode['epoch'] or time.time(),
-                "url": self.urls['addr'] + aNode['addr'],
+                "url": aKnown['url'],
                 "webUrl": self.urls['addr'] + aNode['addr'], # REMOVE
                 "value": float(aNode['balance'] or 0)/Satoshi(),
                 "img": self.urls['img'],
@@ -179,16 +187,17 @@ class CoinController:
             if aNode['label'] != aNode['addr']:
                 aNode['title'] = "Name: {}<br>".format(aNode['label']) + aNode['title']
                 
-            bNode = nodes['to']
+            bNode  = nodes['to']
+            bKnown = self.getKnown(bNode['addr'])
             bNode = {
                 "id": bNode.id,
-                "label": bNode['name'] or bNode['addr'],
+                "label": bKnown['name'],
                 "addr": bNode['addr'],
                 "balance": float(bNode['balance'] or 0),
                 "balVal": float(bNode['balance'] or 0),
                 "group": bNode['wallet'] or 'usdt',
                 "lastUpdate": bNode['epoch'] or time.time(),
-                "url": self.urls['addr'] + bNode['addr'],
+                "url": bKnown['url'],
                 "webUrl": self.urls['addr'] + bNode['addr'], # REMOVE
                 "value": float(bNode['balance'] or 0)/Satoshi(),
                 "img": self.urls['img'],
@@ -207,6 +216,8 @@ class CoinController:
                 "value": float(rel['amount'] or 0),
                 "source": aNode['label'],
                 "target": bNode['label'],
+                "sourceAddr": aNode['addr'],
+                "targetAddr": bNode['addr'],
                 "amount": float(rel['amount'] or 0),
                 "txsNum": int(rel['txsNum'] or 1.0),
                 "avgTx": float(rel['avgTxAmt'] or rel['amount'] or 0),
@@ -239,38 +250,77 @@ class CoinController:
         return data
 
     def getKnown(self, addr):
-        for known in self.getSortedKnown()[0]['addrs']:
+        for known in self.getKnownList()[0]['addrs']:
             if known['addr'] == addr:
                 return known
         
-        return {"name": addr, "addr": addr}
+        return {"name": addr, "addr": addr, "url": self.urls['addr'] + addr}
+    
+    def getKnownDict(self):
+        return json.load(open(self.path + self.coin + ".json"))
 
-    def getSortedKnown(self):
-        nodes = json.load(open(self.path + self.coin + ".json"))
-        categories = [{
-            'category': 'Home',
-            'url': "/{}/search/0".format(self.coin),
-            'addrs': []
-        }]
-        for node in nodes:
-            categories[0]['addrs'].append({
-                'name': node['name'],
-                'url': "/{}/search/{}".format(self.coin, node['addr']),
-                'addr': node['addr']
+    def getKnownList(self):
+        catDict = self.getKnownDict()
+
+        catList    = []
+
+        for cat, info in catDict.items():
+            if cat == "Home":
+                continue
+
+            for addr in info['addrs']:
+                if "Home" not in catDict:
+                    catDict['Home'] = {
+                        'url': "/{}/search/Home?addr[]={}".format(self.coin, addr['addr']),
+                        'addrs': []
+                    }
+                else:
+                    catDict['Home']['url'] += "&addr[]={}".format(addr['addr'])
+
+            catDict['Home']['addrs'] += info['addrs']
+
+            catList.append({
+                'category': cat,
+                'url': info['url'],
+                'addrs': info['addrs']
             })
 
-        return categories
+        if "Home" not in catDict:
+            catList.append({
+                'category': 'Home',
+                'url': "",
+                'addrs': []
+            })
+        else:
+            catList.insert(0, {
+                'category': 'Home',
+                'url': catDict['Home']['url'],
+                'addrs': catDict['Home']['addrs']
+            })
+
+        return catList
     
-    def addAddr(self, addr, name):
+    def addAddr(self, addr, name, cat):
         try: 
-            if self.isValidName(name) and self.isValidAddr(addr) and not self.nameExists(name):
-                addrs = self.getSortedKnown()[0]['addrs'] or []
+            if self.isValidName(name) and self.isValidAddr(addr) and not self.nameExists(name) and self.isValidName(cat):
+                catDict = self.getKnownDict()
                 with open(self.path + self.coin + ".json", "w") as f:
-                    addrs.append({
-                        "name": name,
-                        "addr": addr
-                    })
-                    json.dump(addrs, f)
+                    node = {
+                            "name": name,
+                            "addr": addr,
+                            "url": "/{}/search/{}".format(self.coin, addr)
+                        }
+
+                    if cat in catDict:
+                        catDict[cat]['url'] += "&addr[]={}".format(addr)
+                        catDict[cat]['addrs'].append(node)
+                    else:
+                        catDict[cat] = {
+                            'url': "/{}/search/Home?addr[]={}".format(self.coin, addr),
+                            'addrs': [node]
+                        }
+
+                    json.dump(catDict, f)
 
                     return "Success"
         except Exception as e:
@@ -278,12 +328,15 @@ class CoinController:
 
         return "ERROR"
         
-    def getAddrsText(self, addr, extras=[]):
+    def getAddrsText(self, addr, extras=[], rel=True):
         addrsText = ""
         for x in range(len(addr)):
             string = ""
 
             string += "a.addr = '{}'".format(addr[x])
+            if rel:
+                string = "({} or b.addr = '{}')".format(string, addr[x])
+            
 
             if x < len(extras):
                 string = "({} AND b.addr = '{}')".format(string, extras[x])
@@ -293,19 +346,19 @@ class CoinController:
 
             addrsText += string
 
-        return addrsText
+        return "({})".format(addrsText)
 
     def delAddr(self, addr): 
         try: 
             if self.isValidAddr(addr):
-                addrs = self.getSortedKnown()[0]['addrs'] or []
+                catDict = self.getKnownDict()
                 with open(self.path + self.coin + ".json", "w") as f:
-                    
-                    for i in range(len(addr)):
-                        if addrs[i]['addr'] == addr:
-                            del addrs[i]
-                            json.dump(addrs, f)
-                            return "Success"
+                    for cat, info in catDict.items():
+                        for i in range(len(info['addrs'])):
+                            if catDict[cat]['addrs'][i] == addr:
+                                del catDict[cat]['addrs'][i]
+                                json.dump(catDict, f)
+                                return "Success"
         except Exception as e:
             print(e)
             
@@ -315,7 +368,7 @@ class CoinController:
         return len(name) < 16 and len(name) > 2 and re.match(r'^[A-Za-z0-9_-]*$', name) != None
 
     def nameExists(self, name):
-        addrs = self.getSortedKnown()[0]['addrs']
+        addrs = self.getKnownList()[0]['addrs']
         
         for addr in addrs:
             if addr['name'].lower() == name.lower():
@@ -331,17 +384,25 @@ class CoinController:
                 return True
         return False
 
-    def editAddr(self, addr, name):
+    def editCat(self, prevCat, newCat):
+        try:
+            if self.isValidName(prevCat) and self.isValidName(newCat) and prevCat.lower() != "home":
+                catDict = self.getKnownDict()  
+                if prevCat in catDict:
+                    with open(self.path + self.coin + ".json", "w") as f:
+                        catDict[newCat] = catDict.pop(prevCat)
+                        json.dump(catDict, f)
+                        return "Success"
+        except Exception as e:
+            print(e)
+
+        return "ERROR"
+
+    def editAddr(self, addr, name, cat):
         try: 
-            if self.isValidAddr(addr) and self.isValidName(name) and not self.nameExists(name):
-                addrs = self.getSortedKnown()[0]['addrs'] or []
-                with open(self.coin + ".json", "w") as f:
-                    
-                    for addr in addrs:
-                        if addr['addr'] == addr:
-                            addr['name'] = name
-                            json.dump(addrs, f)
-                            return "Success"
+            if self.isValidAddr(addr) and self.isValidName(name) and not self.nameExists(name) and self.isValidName(cat):
+                if self.delAddr(addr) == "Success" and self.addAddr(addr, name, cat) == "Success":
+                    return "Success"
         except Exception as e:
             print(e)
